requirements.txt
streamlit
pandas
opencv-python
pillow
reportlab
azure-cognitiveservices-vision-computervision
openpyxl